refactor: 야구게임 패키지 경로 변경. * 도메인별 그룹패키지를 적용하여 분류 및 이동.
