+++ src/main/java/baseball/model/ball/Ball.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.model.ball;
 
 import java.util.Objects;
