+++ src/main/java/baseball/model/count/StrikeCount.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.model.count;
 
 import java.util.Objects;
