+++ src/main/java/baseball/service/BallCreateService.java
@@ -2,8 +2,8 @@ package baseball.service;
 
 import java.util.LinkedHashMap;
 
-import baseball.model.BallType;
-import baseball.model.Game;
+import baseball.model.ball.BallType;
+import baseball.model.game.Game;
 
 public class BallCreateService {
