+++ src/test/java/baseball/model/ball/BallNumberTest.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.model.ball;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.*;
 
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import baseball.model.ball.BallNumber;
 
 class BallNumberTest {
