+++ src/test/java/baseball/model/ball/BallTest.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.model.ball;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.*;
 
@@ -7,6 +7,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import baseball.model.ball.Ball;
+import baseball.model.ball.BallType;
 
 class BallTest {
