+++ src/test/java/baseball/model/count/StrikeCountTest.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.model.count;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.*;
 
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import baseball.model.count.StrikeCount;
 
 class StrikeCountTest {
